Move over readNativeFrames to parse.go
